requirements.txt
kabbes_py_starter
kabbes_dir_ops
kabbes_parent_class